complete test using rspec, factorygirl and mocking omniauth.

+++ app/controllers/photos_controller.rb
@@ -15,26 +15,34 @@ class PhotosController < ApplicationController
   # Like function. Send the vote post request every time even though the image may unliked by this user.
   # Like/unlike status will be syncinzed to 500px server.
   def like
-    @id = params[:id]
-    client = f00_client()
-    client.post("photos/#{@id}/vote?vote=1")
-    like = Like.find_or_create_by(user_id: session[:uid], photo_id: id)
-    respond_to do |format|
-      format.html { redirect_to root_path }
-      format.js {}
-    end
+  	if user_logged_in?
+	    id = params[:id]
+	    client = f00_client()
+	    client.post("photos/#{@id}/vote?vote=1")
+	    like = Like.find_or_create_by(user_id: session[:uid], photo_id: id)
+	    respond_to do |format|
+	      format.html { redirect_to root_path }
+	      format.js {}
+	    end
+	  else
+	  	redirect_to "/auth/500px"
+	  end
   end
 
   # Unlike function. Send the vote post request every time even though the image may unliked by this user.
   # Like/unlike status will be syncinzed to 500px server.
   def unlike
-    id = params[:id]
-    client = f00_client()
-    client.delete("photos/#{@id}/vote?vote=0")
-    Like.where(user_id: session[:uid], photo_id: id).delete_all
-    respond_to do |format|
-      format.html { redirect_to root_path }
-      format.js {}
-    end
+  	if user_logged_in?
+	    id = params[:id]
+	    client = f00_client()
+	    client.delete("photos/#{@id}/vote?vote=0")
+	    Like.where(user_id: session[:uid], photo_id: id).delete_all
+	    respond_to do |format|
+	      format.html { redirect_to root_path }
+	      format.js {}
+	    end
+	  else
+	  	redirect_to "/auth/500px"
+	  end
   end
 end
